Updated Solution
### Since Last Commit
---
---
#### Additions
- N/A
---
#### Changes
- Changed references to Vector2f to Vector2i concerning Player/Enemy Placement
---
#### Removed
- N/A

+++ vsSolutions/shadowOperative_001/shadowOperative_001/LevelManager.cpp
@@ -177,6 +177,6 @@ float LevelManager::getTimeLimit() {
 	return m_BaseTimeLimit * m_TimeModifier;
 
 }
-sf::Vector2f LevelManager::getStartPosition() {
+sf::Vector2i LevelManager::getStartPosition() {
 	return m_StartPosition;
 }
